+++ mobil_flutter/lib/presentation/screens/main_navigation_screen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mobil_flutter/presentation/providers/auth_providers.dart'; // YENİ AUTH PROVIDER
+import 'package:mobil_flutter/presentation/providers/auth_providers.dart';
 import 'package:mobil_flutter/presentation/screens/ayarlar_ekrani_misafir.dart';
 import 'package:mobil_flutter/presentation/screens/harita_ekrani.dart';
 import 'package:mobil_flutter/presentation/screens/kesfet_ekrani.dart';
-import 'package:mobil_flutter/presentation/screens/profil_ekrani.dart';
+import 'package:mobil_flutter/presentation/screens/profil/profil_ekrani.dart';
 import 'package:mobil_flutter/presentation/screens/rotalar_ekrani.dart';
 import 'package:mobil_flutter/l10n/app_localizations.dart';
