+++ mobil_flutter/lib/presentation/widgets/sliver_tab_bar_delegate.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+// TabBar'ı SliverAppBar içinde sabit tutmak için yardımcı sınıf
+class SliverAppBarDelegate extends SliverPersistentHeaderDelegate {
+  SliverAppBarDelegate(this._tabBar);
+  final TabBar _tabBar;
+
+  @override
+  double get minExtent => _tabBar.preferredSize.height;
+  @override
+  double get maxExtent => _tabBar.preferredSize.height;
+
+  @override
+  Widget build(BuildContext context, double shrinkOffset, bool overlapsContent) {
+    return Container(
+      color: Theme.of(context).colorScheme.surface,
+      child: _tabBar,
+    );
+  }
+
+  @override
+  bool shouldRebuild(SliverAppBarDelegate oldDelegate) => false;
+}
